--- src/main/java/io/zipcoder/threedaystodeliver/humanresourceapp/Employee.java
@@ -1,10 +1,0 @@
-package io.zipcoder.threedaystodeliver.humanresourceapp;
-
-import java.util.Date;
-
-public class Employee extends Prospect {
-
-    private Compensation compensation;
-    private Date hiredDate;
-
-}
