--- src/main/java/io/zipcoder/threedaystodeliver/humanresourceapp/Human.java
@@ -1,16 +1,0 @@
-package io.zipcoder.threedaystodeliver.humanresourceapp;
-
-public abstract class Human {
-
-    //fields
-    String name;
-    HrContactInfo hrContactInfo;
-
-
-    //constructors
-
-
-    //methods
-    //setters and getters
-
-}
