--- src/main/java/io/zipcoder/threedaystodeliver/humanresourceapp/Prospect.java
@@ -1,13 +1,0 @@
-package io.zipcoder.threedaystodeliver.humanresourceapp;
-
-import java.util.Date;
-
-public class Prospect extends Human{
-
-    private String resume;
-
-    private Date interviewDate;
-    private double score;
-
-
-}
